pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-autoslug>=1.9.9",
    "django-taggit>=6.1.0",
    "dotenv>=0.9.9",
    "pillow>=12.1.1",
    "psycopg>=3.3.3",
    "requests>=2.32.5",
]
